Made it a new flight panel shows up with the temporary text when we simulate the load of a file.
so to test this just try and load a file with the jfilechoose from the 
file menu, a new flightpanel will be auto populated and setup.
please note, everything in theses is placeholder for now.
next step is a temporary graph.
trying to decide between javafx.scene.chart.XYChart
and the jzy3d graphing options.