Minor changes
removed second trim from calculator class (1 should be enough no?)
This invited my IDE to take it upon itself to change the formating :(
sorry.
also tried making the files load to a new folder in a directory, this
did not work  because of permissions issues.

Added the "FlightPanel.LOAD_RAW" and "FlightPanel.LOAD_FLIGHT"
constructor choices to the rsc_jframe.java class.