Loading now works
Loading can not be done from the rsc_jframe, I have modifed the
FlightPanel class to be able to load from a '.rsc' file.
I have also made it so a .rsc file is saved when loading from a raw
file.

Added place holders for each of the toggle buttons actionPerformed
methods.